+++ packages/pluggableWidgets/datagrid-web/src/Datagrid.editorPreview.tsx
@@ -132,6 +132,7 @@ export function preview(props: DatagridPreviewProps): ReactElement {
             page={0}
             paginationType={props.pagination}
             pageSize={numberOfItems}
+            showPagingButtons={props.showPagingButtons}
             loadMoreButtonCaption={props.loadMoreButtonCaption}
             paging={props.pagination === "buttons"}
             pagingPosition={props.pagingPosition}